minor functions relocation across files
