+++ beacon/blockchain/process_proposal.go
@@ -25,11 +25,110 @@ import (
 	"time"
 
 	payloadtime "github.com/berachain/beacon-kit/beacon/payload-time"
+	"github.com/berachain/beacon-kit/consensus/cometbft/service/encoding"
+	"github.com/berachain/beacon-kit/consensus/types"
 	engineerrors "github.com/berachain/beacon-kit/engine-primitives/errors"
 	"github.com/berachain/beacon-kit/errors"
 	"github.com/berachain/beacon-kit/primitives/math"
 	"github.com/berachain/beacon-kit/primitives/transition"
+	cmtabci "github.com/cometbft/cometbft/abci/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 )
+
+const (
+	// BeaconBlockTxIndex represents the index of the beacon block transaction.
+	// It is the first transaction in the tx list.
+	BeaconBlockTxIndex uint = iota
+	// BlobSidecarsTxIndex represents the index of the blob sidecar transaction.
+	// It follows the beacon block transaction in the tx list.
+	BlobSidecarsTxIndex
+)
+
+func (s *Service[
+	_, _, ConsensusBlockT, BeaconBlockT, _, BeaconBlockHeaderT, _, _, _,
+	_, _, _, GenesisT, ConsensusSidecarsT, BlobSidecarsT, _,
+]) ProcessProposal(
+	ctx sdk.Context,
+	req *cmtabci.ProcessProposalRequest,
+) (*cmtabci.ProcessProposalResponse, error) {
+	// Decode the beacon block.
+	blk, err := encoding.
+		UnmarshalBeaconBlockFromABCIRequest[BeaconBlockT](
+		req,
+		BeaconBlockTxIndex,
+		s.chainSpec.ActiveForkVersionForSlot(math.U64(req.Height)),
+	)
+	if err != nil {
+		return createProcessProposalResponse(errors.WrapNonFatal(err))
+	}
+
+	// Decode the blob sidecars.
+	sidecars, err := encoding.
+		UnmarshalBlobSidecarsFromABCIRequest[BlobSidecarsT](
+		req,
+		BlobSidecarsTxIndex,
+	)
+	if err != nil {
+		return createProcessProposalResponse(errors.WrapNonFatal(err))
+	}
+
+	// Process the blob sidecars, if any
+	if !sidecars.IsNil() && sidecars.Len() > 0 {
+		var consensusSidecars *types.ConsensusSidecars[
+			BlobSidecarsT,
+			BeaconBlockHeaderT,
+		]
+		consensusSidecars = consensusSidecars.New(
+			sidecars,
+			blk.GetHeader(),
+		)
+
+		s.logger.Info("Received incoming blob sidecars")
+
+		// TODO: Clean this up once we remove generics.
+		c := convertConsensusSidecars[
+			ConsensusSidecarsT,
+			BlobSidecarsT,
+			BeaconBlockHeaderT,
+		](consensusSidecars)
+
+		// Verify the blobs and ensure they match the local state.
+		err = s.blobProcessor.VerifySidecars(c)
+		if err != nil {
+			s.logger.Error(
+				"rejecting incoming blob sidecars",
+				"reason", err,
+			)
+			return createProcessProposalResponse(errors.WrapNonFatal(err))
+		}
+
+		s.logger.Info(
+			"Blob sidecars verification succeeded - accepting incoming blob sidecars",
+			"num_blobs",
+			sidecars.Len(),
+		)
+	}
+
+	// Process the block
+	var consensusBlk *types.ConsensusBlock[BeaconBlockT]
+	consensusBlk = consensusBlk.New(
+		blk,
+		req.GetProposerAddress(),
+		req.GetTime(),
+	)
+	err = s.VerifyIncomingBlock(
+		ctx,
+		consensusBlk.GetBeaconBlock(),
+		consensusBlk.GetConsensusTime(),
+		consensusBlk.GetProposerAddress(),
+	)
+	if err != nil {
+		s.logger.Error("failed to verify incoming block", "error", err)
+		return createProcessProposalResponse(errors.WrapNonFatal(err))
+	}
+
+	return createProcessProposalResponse(nil)
+}
 
 // VerifyIncomingBlock verifies the state root of an incoming block
 // and logs the process.
@@ -180,4 +279,31 @@ func (s *Service[
 	_, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _,
 ]) shouldBuildOptimisticPayloads() bool {
 	return s.optimisticPayloadBuilds && s.localBuilder.Enabled()
+}
+
+// createResponse generates the appropriate ProcessProposalResponse based on the
+// error.
+func createProcessProposalResponse(
+	err error,
+) (*cmtabci.ProcessProposalResponse, error) {
+	status := cmtabci.PROCESS_PROPOSAL_STATUS_REJECT
+	if !errors.IsFatal(err) {
+		status = cmtabci.PROCESS_PROPOSAL_STATUS_ACCEPT
+		err = nil
+	}
+	return &cmtabci.ProcessProposalResponse{Status: status}, err
+}
+
+func convertConsensusSidecars[
+	ConsensusSidecarsT any,
+	BlobSidecarsT any,
+	BeaconBlockHeaderT any,
+](
+	cSidecars *types.ConsensusSidecars[BlobSidecarsT, BeaconBlockHeaderT],
+) ConsensusSidecarsT {
+	val, ok := any(cSidecars).(ConsensusSidecarsT)
+	if !ok {
+		panic("failed to convert conesensusSidecars to ConsensusSidecarsT")
+	}
+	return val
 }
